@@ -1,93 +1,132 @@
 {
-  "name": "ts-react",
-  "version": "0.1.0",
-  "private": true,
+  "name": "searchify-client",
+  "version": "0.0.1",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "jest --no-cache",
+    "build": "./node_modules/.bin/webpack --progress --define process.env.NODE_ENV=\"'development'\"",
+    "build-production": "./node_modules/.bin/webpack --env.production -p",
+    "serve": "./node_modules/.bin/webpack-dev-server --progress --define process.env.NODE_ENV=\"'development'\"",
+    "watch": "./node_modules/.bin/webpack --watch",
+    "setup": "yarn install && node ./node_modules/husky/bin/install.js",
+    "postcheckout": "yarn install && exit 0"
+  },
+  "author": "Christian Esswein",
   "dependencies": {
-    "@types/jest": "^19.2.3",
-    "@types/node": "^7.0.18",
+    "@types/classnames": "^2.2.0",
+    "@types/clone": "^0.1.30",
+    "@types/deep-equal": "0.0.30",
     "@types/react": "^15.0.24",
+    "@types/react-addons-shallow-compare": "^0.14.19",
+    "@types/react-autosuggest": "^8.0.1",
     "@types/react-dom": "^15.5.0",
+    "@types/react-hot-loader": "^3.0.1",
     "@types/react-redux": "^4.4.40",
-    "@types/three": "^0.84.8",
+    "@types/react-router": "^3.0.0",
+    "@types/react-router-redux": "^4.0.0",
+    "@types/react-select": "^1.0.42",
+    "@types/redux": "^3.6.0",
+    "@types/redux-form": "^6.6.3",
+    "@types/webpack-env": "^1.13.0",
+    "autoprefixer": "^7.1.1",
+    "awesome-typescript-loader": "^3.1.3",
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.1.0",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-env": "^1.5.2",
+    "classnames": "2.2.5",
+    "clone": "^2.1.1",
+    "compression-webpack-plugin": "^0.4.0",
+    "css-loader": "^0.28.4",
+    "deep-equal": "^1.0.1",
+    "doiuse": "^3.0.0",
+    "html-webpack-harddisk-plugin": "^0.1.0",
+    "html-webpack-plugin": "^2.28.0",
+    "husky": "^0.13.4",
+    "immutable": "4.0.0-rc.2",
+    "jest": "^20.0.4",
+    "jest-fetch-mock": "^1.1.1",
+    "js-uuid": "^0.0.6",
+    "json5-loader": "^1.0.1",
+    "postcss-loader": "^2.0.6",
+    "postcss-partial-import": "^4.1.0",
+    "postcss-scss": "^1.0.1",
+    "postcss-strip-inline-comments": "^0.1.5",
+    "precss": "1.4.0",
     "react": "^15.5.4",
-    "react-autosuggest": "^9.0.1",
     "react-dom": "^15.5.4",
+    "react-highlight": "^0.9.0",
+    "react-hot-loader": "3.0.0-beta.7",
     "react-redux": "^5.0.5",
-    "react-select": "^1.0.0-rc.5",
-    "redux": "^3.6.0",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.2.0",
+    "react-router": "^3.0.0",
+    "react-router-redux": "^4.0.8",
+    "react-select": "^1.0.0-rc.3",
+    "redux": "^3.7.0",
+    "redux-form": "^6.8.0",
+    "redux-thunk": "^2.1.0",
+    "style-loader": "^0.18.2",
+    "to-string-loader": "^1.1.5",
+    "ts-jest": "^20.0.6",
+    "tslint": "^5.4.3",
+    "tslint-eslint-rules": "^4.1.1",
+    "typescript": "^2.3.4",
+    "webpack": "^3.0.0",
+    "webpack-dev-server": "^2.5.0",
+    "react-autosuggest": "^9.0.1",
     "three": "^0.85.2",
     "three-trackballcontrols": "0.0.5"
   },
   "devDependencies": {
-    "app-root-path": "^2.0.1",
-    "autoprefixer": "6.7.2",
-    "case-sensitive-paths-webpack-plugin": "1.1.4",
-    "chalk": "1.1.3",
-    "cli-highlight": "1.1.4",
-    "connect-history-api-fallback": "1.3.0",
-    "cross-spawn": "4.0.2",
-    "css-loader": "0.26.1",
-    "detect-port": "1.1.0",
-    "dotenv": "2.0.0",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.10.0",
-    "fs-extra": "0.30.0",
-    "html-webpack-plugin": "2.24.0",
-    "http-proxy-middleware": "0.17.3",
-    "jest": "18.1.0",
-    "json-loader": "0.5.4",
-    "object-assign": "4.1.1",
-    "postcss-loader": "1.2.2",
-    "promise": "7.1.1",
-    "react-dev-utils": "^0.5.2",
-    "style-loader": "0.13.1",
-    "ts-loader": "^1.3.1",
-    "tslint": "^4.0.2",
-    "tslint-loader": "^3.3.0",
-    "tslint-react": "^2.0.0",
-    "typescript": "^2.3.2",
-    "url-loader": "0.5.7",
-    "webpack": "1.14.0",
-    "webpack-dev-server": "1.16.2",
-    "webpack-manifest-plugin": "1.1.0",
-    "whatwg-fetch": "2.0.2"
+    "@types/jest": "^20.0.1",
+    "extract-text-webpack-plugin": "^2.1.2",
+    "postcss-import": "^10.0.0"
   },
-  "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js --env=jsdom"
-  },
-  "proxy": "http://localhost:5000",
+  "browserslist": [
+    "ie >= 11",
+    "last 2 Chrome versions",
+    "last 2 Firefox versions",
+    "last 2 Edge versions",
+    "Firefox ESR"
+  ],
   "jest": {
+    "verbose": true,
+    "browser": true,
+    "collectCoverage": true,
+    "mapCoverage": true,
+    "coverageReporters": [
+      "cobertura"
+    ],
     "collectCoverageFrom": [
-      "src/**/*.{ts,tsx}"
+      "src/**/*.{js,jsx,ts,tsx}",
+      "!**/node_modules/**"
     ],
-    "setupFiles": [
-      "<rootDir>/config/polyfills.js"
+    "setupTestFrameworkScriptFile": "<rootDir>/env-config/setup-jasmine-env.js",
+    "unmockedModulePathPatterns": [
+      "./node_modules/jasmine-reporters"
     ],
-    "testPathIgnorePatterns": [
-      "<rootDir>[/\\\\](build|docs|node_modules|scripts)[/\\\\]"
-    ],
-    "testEnvironment": "node",
-    "testURL": "http://localhost",
-    "transform": {
-      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-      "^.+\\.tsx?$": "<rootDir>/config/jest/typescriptTransform.js",
-      "^(?!.*\\.(css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-    },
-    "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx)$"
-    ],
-    "moduleNameMapper": {
-      "^react-native$": "react-native-web"
-    },
     "moduleFileExtensions": [
+      "js",
+      "jsx",
+      "json",
       "ts",
-      "tsx",
-      "js"
+      "tsx"
     ],
-    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$"
+    "notify": true,
+    "roots": [
+      "src/"
+    ],
+    "testRegex": ".*\\.test\\.(ts|tsx|js)$",
+    "transform": {
+      ".(ts|tsx)": "<rootDir>/env-config/jest-preprocessor.js"
+    },
+    "modulePaths": [
+      "src",
+      "assets"
+    ],
+    "automock": false,
+    "setupFiles": [
+      "./env-config/jest-setup.js"
+    ]
   }
 }
