@@ -1,33 +1,22 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import App from './containers/app';
-
-import './index.css';
+import { ConnectedApp } from 'components/app';
 
 import { createStore, applyMiddleware } from 'redux';
-import { items } from './reducers/index';
-import { StoreState } from './types/index';
+import { items } from 'reducers/index';
+import { StoreState } from 'types/index';
 import { Provider } from 'react-redux';
 import thunkMiddleware from 'redux-thunk';
-import { DispatchContextProvider } from './containers/dispatchContextProvider';
-const logger = require('redux-logger');
+import { DispatchContextProvider } from 'components/dispatchContextProvider';
+import { AppContainer as ReactHotLoaderAppContainer } from 'react-hot-loader';
 
-declare var module: { hot: { accept: (path?: string, callback?: () => void) => void } };
+require('./index.css');
 
-function configureStore(initialState: StoreState): any { // Store<StoreState>
-  const store = createStore<StoreState>(items, initialState, applyMiddleware(thunkMiddleware, logger.createLogger()));
-
-  if (module.hot) {
-    module.hot.accept();
-    // Enable Webpack hot module replacement for reducers
-    module.hot.accept('./reducers', () => {
-      const nextRootReducer = require('./reducers').default;
-      store.replaceReducer(nextRootReducer);
-    });
-  }
-
+function configureStore(initialState: StoreState) {
+  const store = createStore(items as any, initialState, applyMiddleware(thunkMiddleware));
   return store;
 }
+
 
 const store = configureStore({
     queryItems: [],
@@ -36,11 +25,37 @@ const store = configureStore({
   }
 );
 
-ReactDOM.render(
-  <Provider store={store}>
-    <DispatchContextProvider dispatch={store.dispatch}>
-      <App />
-    </DispatchContextProvider>
-  </Provider>,
-  document.getElementById('root') as HTMLElement
-);
+
+const render = (App: React.ComponentClass<{}>) => {
+  ReactDOM.render(
+    (
+      <ReactHotLoaderAppContainer>
+
+        <DispatchContextProvider dispatch={store.dispatch}>
+          <Provider store={store}>
+            <App />
+          </Provider>
+        </DispatchContextProvider>
+
+      </ReactHotLoaderAppContainer>
+    ),
+    document.getElementById('root'),
+  );
+};
+
+render(ConnectedApp);
+
+
+// Hot Module Replacement API https://github.com/gaearon/react-hot-loader/tree/master/docs#migration-to-30
+declare const module: any;
+if (module.hot) {
+  module.hot.accept(
+    'components/app',
+    () => {
+      render(ConnectedApp);
+
+      console.log('react hot reloaded');
+    },
+  );
+}
+
